12:09 26/12/2023

+++ DatBanNhaHang/Services/Implements/TrangThaiHoaDonServices.cs
@@ -21,20 +21,23 @@ namespace DatBanNhaHang.Services.Implements
         public async Task<PageResult<TrangThaiHoaDonDTOs>> HienThiTrangThaiHoaDon(int id, int pageSize, int pageNumber)
         {
             var lstMonAn = id == 0 ? 
-                contextDB.TrangThaiHoaDon.Where(y=>y.status==1).Select(x => converters.EntityToDTOs(x)) 
-                : contextDB.TrangThaiHoaDon.Where(y => y.id == id&& y.status==1).Select(x => converters.EntityToDTOs(x));
+                contextDB.TrangThaiHoaDon
+                    .Where(y=>y.status==1)
+                    .Select(x => converters.EntityToDTOs(x)) 
+                : contextDB.TrangThaiHoaDon
+                    .Where(y => y.id == id&& y.status==1)
+                    .Select(x => converters.EntityToDTOs(x));
             var result = Pagintation.GetPagedData(lstMonAn, pageSize, pageNumber);
             return result;
         }
-
         public async Task<ResponseObject<TrangThaiHoaDonDTOs>> SuaTrangThaiHoaDon(int id, Request_SuaTrangThaiHoaDon request)
         {
-            var tthd = contextDB.TrangThaiHoaDon.SingleOrDefault(x => x.id == id);
+            var tthd = contextDB.TrangThaiHoaDon.SingleOrDefault(x => x.id == id && x.status==1);
             if (tthd == null)
             {
                 return response.ResponseError(StatusCodes.Status404NotFound, "Không tồn tại trạng thái này ", null);
             }
-            tthd.TenTrangThai = request.TenTrangThai == null ? tthd.TenTrangThai : request.TenTrangThai;
+            tthd.TenTrangThai = request.TenTrangThai ?? tthd.TenTrangThai ;
             contextDB.Update(tthd);
             await contextDB.SaveChangesAsync();
             return response.ResponseSuccess("Sửa trạng thái thành công ", converters.EntityToDTOs(tthd));
@@ -64,13 +67,16 @@ namespace DatBanNhaHang.Services.Implements
 
             tthd.status = 2;
             var lsthd = contextDB.HoaDon.Where(x => x.id == tthd.id).ToList();
-            foreach (var hd in lsthd)
+            if (lsthd.Count>0)
             {
-                hd.status = 2;
-                /*var lstCT = contextDB.ChiTietHoaDon.Where(x => x.HoaDonID == hd.id);
-                contextDB.RemoveRange(lstCT);*/
+                foreach (var hd in lsthd)
+                {
+                    hd.status = 2;
+                    /*var lstCT = contextDB.ChiTietHoaDon.Where(x => x.HoaDonID == hd.id);
+                    contextDB.RemoveRange(lstCT);*/
+                }
+                contextDB.UpdateRange(lsthd);
             }
-            contextDB.Update(lsthd);
             contextDB.Update(tthd);
             await contextDB.SaveChangesAsync();
             return response.ResponseSuccess("Xoá trạng thái thành công ", converters.EntityToDTOs(tthd));
